# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/feature-flags": {
      "get": {
        "tags": [
          "Internal"
        ],
        "summary": "List all feature flags (global and per-org). Admin only.",
        "operationId": "list_all",
        "responses": {
          "200": {
            "description": "All feature flags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureFlagItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/internal/feature-flags/{name}": {
      "put": {
        "tags": [
          "Internal"
        ],
        "summary": "Set a global or per-org feature flag. Admin only.",
        "operationId": "set_flag",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Flag name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "org_id",
            "in": "query",
            "description": "Org UUID for per-org override",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetFlagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flag updated",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/orgs/{org_id}/feature-flags": {
      "get": {
        "tags": [
          "Feature Flags"
        ],
        "summary": "List effective feature flags for an org (global + org overrides). Org member only.",
        "operationId": "list_for_org",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Org UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Effective flags for org",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureFlagEffectiveItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeatureFlagEffectiveItem": {
        "type": "object",
        "required": [
          "name",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "FeatureFlagItem": {
        "type": "object",
        "required": [
          "name",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "org_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SetFlagRequest": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}